requirements.txt
langchain==0.1.14
python-dotenv==1.0.1
unstructured[local-inference]==0.5.6
chromadb==0.4.24 # Vector storage
openai==1.16.2 # For embeddings
tiktoken==0.6.0  # For embeddings
pymupdf==1.24.1
rapidocr-onnxruntime==1.3.15
streamlit==1.33.0
pypdf==4.1.0